+++ test/feedback.py
@@ -116,4 +116,4 @@ def test_strategy():
 def test_axioms():
     from discopy import testing
 
-    testing.assert_axioms(Ty, Diagram, Functor)
+    testing.assert_axioms(Ty, Diagram, Hypergraph, Functor)