+++ test/traced.py
@@ -48,4 +48,4 @@ def test_strategy():
 def test_axioms():
     from discopy import testing
 
-    testing.assert_axioms(Diagram, Functor)
+    testing.assert_axioms(Diagram, Hypergraph, CMap, Functor)